Json (de)Serialization of ulong is now forced to be string for javascript compatiblity

+++ src/Coinecta.API/Services/TransactionBuildingService.cs
@@ -83,7 +83,7 @@ public class TransactionBuildingService(IDbContextFactory<CoinectaDbContext> dbC
         {
             Token = new()
             {
-                { stakePoolData.StakePool.AssetName, request.Amount }
+                { stakePoolData.StakePool.AssetName, (long)request.Amount }
             }
         });
